AUI-1602 Resize Iframe does not detect height of modal contents
This component is supposed to resize to show all of its contents.
Previously however, if there were modal elements in the contents of the
iframe their height would not be detected and so they would appear cutoff.
This fix finds height of the largest modal item and then will use this
value as the height for the iframe if it is larger than the height of the
other contents.